backend/requirements.txt
apscheduler
requests
chromadb
openai
python-dotenv
google-genai
